refactor: service와 usecase 패키지를 application 안으로 이동
