+++ src/main/java/racingcar/application/service/RacingCarGameService.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.application.service;
 
 import racingcar.domain.car.Car;
 import racingcar.domain.race.RaceChecker;
