+++ src/test/java/racingcar/application/service/CarServiceTest.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.application.service;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,7 +8,6 @@ import racingcar.domain.util.NumberGenerator;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class CarServiceTest {
     NumberGenerator dummyGenerator = () -> 0;
