+++ src/test/java/racingcar/application/service/RacingCarGameServiceTest.java
@@ -1,4 +1,4 @@
-package racingcar.service;
+package racingcar.application.service;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
